+++ app/security/page.tsx
@@ -1,0 +1,15 @@
+import { StaticPage } from "@/components/StaticPage";
+import SecurityMd from "./security.mdx";
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Security",
+};
+
+export default function Page() {
+  return (
+    <StaticPage>
+      <SecurityMd />
+    </StaticPage>
+  );
+}
