+++ app/security/security.mdx
@@ -1,0 +1,9 @@
+# Security
+
+## Report
+
+If you would like to report a vulnerability, or have any security concerns with a Argos product, please e-mail [contact@argos-ci.com](mailto:contact@argos-ci.com).
+
+When disclosing a vulnerability, please include a proof of concept, a list of any tools used (including versions), and the output of any tools used.
+
+We take all disclosures very seriously. Once disclosures are receivied, we rapidly respond and verify each vulnerability before taking the necessary steps to fix them. Once verified, we periodically send status updates as problems are fixed.
