+++ components/Container.tsx
@@ -1,12 +1,11 @@
 import { clsx } from "clsx";
 
-export const Container = ({
-  className,
-  ...props
-}: React.ComponentProps<"div">) => {
+export type ContainerProps = React.ComponentProps<"div">;
+
+export const Container = ({ className, ...props }: ContainerProps) => {
   return (
     <div
-      className={clsx(className, "container mx-auto px-4 sm:px-8 max-w-5xl")}
+      className={clsx("container mx-auto px-4 sm:px-8 max-w-5xl", className)}
       {...props}
     />
   );
