+++ components/StaticPage.tsx
@@ -1,19 +1,12 @@
 import { Container } from "@/components/Container";
-import { Head } from "@/components/Head";
 
-export const StaticPage: React.FC<{
-  children: React.ReactNode;
-  title: string;
-}> = ({ children, title }) => {
+export const StaticPage = ({ children }: { children: React.ReactNode }) => {
   return (
-    <>
-      <Head title={title} />
-      <article
-        className="prose prose-invert mx-auto max-w-none mt-14 mb-24"
-        style={{ contain: "none" }}
-      >
-        <Container>{children}</Container>
-      </article>
-    </>
+    <article
+      className="prose prose-invert mx-auto max-w-none mt-14 mb-24"
+      style={{ contain: "none" }}
+    >
+      <Container>{children}</Container>
+    </article>
   );
 };
