+++ components/Tooltip.tsx
@@ -1,4 +1,5 @@
-import React from "react";
+"use client";
+import * as React from "react";
 import * as TooltipPrimitive from "@radix-ui/react-tooltip";
 
 export type TooltipProps = {
