+++ components/useLiveRef.ts
@@ -1,0 +1,17 @@
+import * as React from "react";
+
+/**
+ * Creates a `React.RefObject` that is constantly updated with the incoming
+ * value.
+ * @example
+ * function Component({ prop }) {
+ *   const propRef = useLiveRef(prop);
+ * }
+ */
+export function useLiveRef<T>(value: T) {
+  const ref = React.useRef(value);
+  React.useLayoutEffect(() => {
+    ref.current = value;
+  });
+  return ref;
+}
