+++ package.json
@@ -20,11 +20,13 @@
     "@radix-ui/react-dialog": "^1.0.4",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-tooltip": "^1.0.6",
+    "@types/mdx": "^2.0.5",
     "clsx": "^1.2.1",
     "fast-glob": "^3.3.1",
     "gray-matter": "^4.0.3",
     "next": "^13.4.12",
     "next-mdx-remote": "^4.4.1",
+    "next-plausible": "^3.10.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-syntax-highlighter": "^15.5.0",
