--- pages/_document.tsx
@@ -1,21 +1,0 @@
-import Document, { Html, Head, Main, NextScript } from "next/document";
-
-export default class MyDocument extends Document {
-  render() {
-    return (
-      <Html>
-        <Head>
-          <script
-            defer
-            data-domain="argos-ci.com"
-            src="https://plausible.io/js/script.js"
-          ></script>
-        </Head>
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    );
-  }
-}
