--- pages/privacy.tsx
@@ -1,10 +1,0 @@
-import { StaticPage } from "@/components/StaticPage";
-import PrivacyMd from "@/markdown/privacy.md";
-
-export default function Privacy() {
-  return (
-    <StaticPage title="Privacy">
-      <PrivacyMd />
-    </StaticPage>
-  );
-}
