--- pages/security.tsx
@@ -1,10 +1,0 @@
-import { StaticPage } from "@/components/StaticPage";
-import SecurityMd from "@/markdown/security.md";
-
-export default function Privacy() {
-  return (
-    <StaticPage title="Security">
-      <SecurityMd />
-    </StaticPage>
-  );
-}
