--- pages/terms.tsx
@@ -1,10 +1,0 @@
-import { StaticPage } from "@/components/StaticPage";
-import TermsMd from "@/markdown/terms.md";
-
-export default function Terms() {
-  return (
-    <StaticPage title="Terms">
-      <TermsMd />
-    </StaticPage>
-  );
-}
